{"schema":"decodex/commit/1","summary":"Move committed-only downward recovery","authority":"manual"}
